refactor: remove pkg folder from api server
