+++ api/server/asset_scan_estimation_controller.go
@@ -23,8 +23,8 @@ import (
 	"github.com/labstack/echo/v4"
 	"gorm.io/gorm"
 
-	"github.com/openclarity/vmclarity/api/server/pkg/common"
-	dbtypes "github.com/openclarity/vmclarity/api/server/pkg/database/types"
+	"github.com/openclarity/vmclarity/api/server/common"
+	dbtypes "github.com/openclarity/vmclarity/api/server/database/types"
 	"github.com/openclarity/vmclarity/api/types"
 	"github.com/openclarity/vmclarity/core/to"
 )
