+++ api/server/database/gorm/database.go
@@ -25,8 +25,8 @@ import (
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
 
-	"github.com/openclarity/vmclarity/api/server/pkg/database/odatasql/jsonsql"
-	"github.com/openclarity/vmclarity/api/server/pkg/database/types"
+	"github.com/openclarity/vmclarity/api/server/database/odatasql/jsonsql"
+	"github.com/openclarity/vmclarity/api/server/database/types"
 )
 
 func NewDatabase(config types.DBConfig) (types.Database, error) {
