+++ api/server/database/gorm/odata.go
@@ -22,8 +22,8 @@ import (
 	"gorm.io/datatypes"
 	"gorm.io/gorm"
 
-	"github.com/openclarity/vmclarity/api/server/pkg/database/odatasql"
-	"github.com/openclarity/vmclarity/api/server/pkg/database/odatasql/jsonsql"
+	"github.com/openclarity/vmclarity/api/server/database/odatasql"
+	"github.com/openclarity/vmclarity/api/server/database/odatasql/jsonsql"
 	"github.com/openclarity/vmclarity/core/to"
 )
