+++ api/server/database/odatasql/jsonsql/variant.go
