+++ api/server/database/odatasql/query_test.go
@@ -30,7 +30,7 @@ import (
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
 
-	"github.com/openclarity/vmclarity/api/server/pkg/database/odatasql/jsonsql"
+	"github.com/openclarity/vmclarity/api/server/database/odatasql/jsonsql"
 )
 
 type SubOption struct {
